% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data.frame.R
\name{as_data.frame}
\alias{as_data.frame}
\alias{as_data.frame.specimens}
\alias{as_data.frame,specimens-method}
\title{Coerce specimens to data frame objects}
\usage{
as_data.frame(x, ...)

\method{as_data.frame}{specimens}(x, ...)
}
\arguments{
\item{x}{A \linkS4class{specimens}.}

\item{...}{Further arguments passed among methods (not used yet).}
}
\value{
A \link{data.frame} object.
}
\description{
A straight-forward method to coerce objects of class \linkS4class{specimens} to
\link{data.frame}. In the case that the collection is an object of class \link{sf},
coordinate values will be extracted as \strong{longitude} and \strong{latitude}.

In the case of multiple determinations in the history of a specimen only the
newest one will be included in the data frame.
}
