% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churo_survey-data.R
\docType{data}
\name{churo_survey-data}
\alias{churo_survey-data}
\alias{churo_survey}
\title{List of specimens collected in the Kenyan Rift Valley}
\description{
A description will follow...
}
\keyword{data}
