% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\alias{clean,specimens-method}
\title{Delete orphaned records in specimens}
\usage{
clean(object, ...)

\S4method{clean}{specimens}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{specimens} object.}

\item{...}{Further arguments passed from or to other methods.}
}
\value{
A clean \linkS4class{specimens} object.
}
\description{
Manipulation of slots may generate orphaned entries in \linkS4class{specimens}
objects, thus such information needs to be cleaned to make objects valid
again.
}
\author{
Miguel Alvarez \email{kamapu@posteo.de}
}
