% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{coerce-methods}
\alias{coerce-methods}
\alias{coerce,sf,data.frame-method}
\alias{coerce,specimens,data.frame-method}
\title{Coerce specimens to data frame objects}
\arguments{
\item{from}{A \linkS4class{specimens} or a sf object.}
}
\value{
A \link{data.frame} object.
}
\description{
A straight-forward method to coerce objects of class \linkS4class{specimens} to
\link{data.frame}. In the case that the collection is an object of class \link{sf},
coordinate values will be extracted as \strong{coord_x} and \strong{coord_y}.

In the case of multiple determinations in the history of a specimen only the
newest one will be included in the data frame.
}
\examples{
# Coerce sf to data.frame
as(churo_survey@collections, "data.frame")

# Coerce specimens to data.frame
as(churo_survey, "data.frame")
}
