% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_specimens.R
\name{map_specimens}
\alias{map_specimens}
\alias{map_specimens.specimens}
\alias{map_specimens,specimens-method}
\title{Display location of specimens in a leaflet widget}
\usage{
map_specimens(x, ...)

\method{map_specimens}{specimens}(
  x,
  add_cols = c("coll_date", "leg", "taxon_name"),
  sep = " | ",
  date_format = "\%d.\%m.\%Y",
  p_tiles = "OpenStreetMap",
  ...
)
}
\arguments{
\item{x}{A \linkS4class{specimens} object.}

\item{...}{Further arguments passed among methods.}

\item{add_cols}{A character vector with the names of columns to be displayed
in the popup labels of markers.}

\item{sep}{A character value used as separator between values of different
columns in the popup labels.}

\item{date_format}{A character value indicating the format used for the
collection date. It is passed to the function \code{\link[=format.Date]{format.Date()}}.}

\item{p_tiles}{Character value passed to \code{\link[=addProviderTiles]{addProviderTiles()}}. The tiles
used as background in the map.}
}
\description{
Display of specimens' location in a leaflet widget. Markers sharing locations
will be clustered.
}
\examples{
\dontrun{
map_specimens(churo_survey)
}

}
