% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimens-class.R
\docType{class}
\name{specimens-class}
\alias{specimens-class}
\title{An S4 class containing specimen collections.}
\description{
A list of collections of specimens considering duplicated specimens and
history of determination.
This package was targeting collection databases from the perspective of a
collector and designed for collections of herbarium specimens but also
applicable to collections of other organisms, for instance zoological
collections as well.
}
\section{Slots}{

\describe{
\item{\code{collections}}{A data frame or an object of class \link[sf:sf]{sf} including
variables associated to a collection, meaning a sample of a species in a
locality. Mandatory columns are here \strong{coll_nr} (unique identifier
of the collection), \strong{leg} (name of collector), and \strong{coll_date}
(date of collection set as class \link[=as.Date]{Date}).}

\item{\code{specimens}}{A data frame with information specifical to single specimens.
It may be one specimen per collection or many (duplicates). This database
list can contain the final destination of the specimens, in the case of
plant specimens, the herbarium. Mandatory columns in this table are
\strong{spec_id} (unique identifier for specimens) and \strong{coll_nr}
(a foreign key pointing to the homonimous column in slot 'collections').}

\item{\code{history}}{A data frame containing the history of determination. In the
case of herbarium specimens, the update of the species determination.
Mandatory columns are \strong{fid} (unique identifier), \strong{spec_id}
(foreign key pointing to homonimous columns in slot 'specimens'),
\strong{det} (the name of determining person), \strong{det_date}
(date of update set as class \link[=as.Date]{Date}), and \strong{taxon_name}
(the scientific name of specimen according to the determinator).}
}}

\examples{
showClass("specimens")

## Create an empty object
spec <- new("specimens")
}
\author{
Miguel Alvarez \email{kamapu@posteo.de}
}
