% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\alias{subset,specimens-method}
\title{Subset method for specimens objects}
\usage{
\S4method{subset}{specimens}(x, subset, slot = "collections", ...)
}
\arguments{
\item{x}{Object of class \linkS4class{specimens}.}

\item{subset}{Logical vector or logical operation to apply as subset.}

\item{slot}{Character value indicating the slot to be used for the subset.
This is not applied to slot 'history'.}

\item{...}{Further arguments to be passed to or from other methods.}
}
\value{
An object of class \linkS4class{specimens}.
}
\description{
Creating subsets of \linkS4class{specimens} objects by using logical operations.
}
\author{
Miguel Alvarez \email{kamapu@posteo.de}
}
