% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary,specimens-method}
\alias{show,specimens-method}
\alias{print,specimens-method}
\alias{print}
\title{Print overviews for 'specimens'}
\usage{
\S4method{summary}{specimens}(object, coll_nr = "all", spec_id, format = "\%d.\%m.\%Y", ...)

\S4method{show}{specimens}(object)

\S4method{print}{specimens}(x, ...)
}
\arguments{
\item{object, x}{A \linkS4class{specimens} object.}

\item{coll_nr}{Vector with collections IDs to be displayed in the summary.
The keyword "all" is set as default indicating that all collections will
be included in the summary.}

\item{spec_id}{Vector with IDs of specimens that will result in a display of
the determination history of the specimens but will be ignored if values
for 'coll_nr' are provided.}

\item{format}{Character value indicating the format applied to dates in the
displays. See \code{\link[=strptime]{strptime()}} for alternative formats.}

\item{...}{Further arguments passed to or from another methods.}
}
\description{
A method to display either an overview of the content of
\linkS4class{specimens} objects or an overview of selected taxa.
}
\author{
Miguel Alvarez \email{kamapu@posteo.de}
}
