% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{view_collection}
\alias{view_collection}
\title{Function producing the overview per collection.}
\usage{
view_collection(object, format = "\%d.\%m.\%Y")
}
\description{
Function producing the overview per collection.
}
\keyword{internal}
