% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{view_specimens}
\alias{view_specimens}
\title{Function producing the overview of whole object.}
\usage{
view_specimens(object)
}
\description{
Function producing the overview of whole object.
}
\keyword{internal}
