% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_table.R
\name{view_table}
\alias{view_table}
\alias{view_table.specimens}
\alias{view_table,specimens-method}
\title{Display an interactive table using shiny}
\usage{
view_table(x, ...)

\method{view_table}{specimens}(
  x,
  add_cols = c("coll_date", "leg", "taxon_name"),
  date_format = "\%d.\%m.\%Y",
  ...
)
}
\arguments{
\item{x}{A \linkS4class{specimens} object.}

\item{...}{Further arguments passed among methods.}

\item{add_cols}{A character vector with the names of columns to be displayed
in the table.}

\item{date_format}{A character value indicating the format used for the
collection date. It is passed to the function \code{\link[=format.Date]{format.Date()}}.}
}
\description{
Table with specimens in an interactive session.

Note that the shiny app will freeze your R session as long as the app is
running. The app will stop if you close the respective tab.
}
\examples{
\dontrun{
view_table(churo_survey)
}

}
